+++ static/scripts/modules/blocklyHandling.js
@@ -4,8 +4,8 @@ import * as Blockly from "blockly";
 import beautify from "js-beautify";
 import { clearLog, handleLogFromWorker } from "./logging";
 import "regenerator-runtime/runtime";
-import "../newblocks";
-import "../threadblocks";
+import "../normalBlockBehaviour";
+import "../threadBlockBehaviour";
 import { addPrintOutput } from "../workspace";
 import { updateValue as updateValuePlot, drawPlots } from "../PlotHandler";
 import { updateValue as updateValueCSV, printDoneMessage } from "../CSVHandler";
