+++ static/scripts/normalBlockBehaviour.js
