+++ static/scripts/threadBlockBehaviour.js
